optimize ProcessClass structure and test python

+++ BYSJ_GUI/test.py
@@ -1,15 +1,25 @@
 import cv2
 import argparse
 
-ap = argparse.ArgumentParser(description="test")
-ap.add_argument("-i","--input",type=str,help="input path")
+ap = argparse.ArgumentParser(description="python part")
+input = ap.add_mutually_exclusive_group()
+input.add_argument("-f","--file",type=str,help="input file path",default="default")
+input.add_argument("--folder",type=str,help="input folder path",default="default")
 arg = ap.parse_args()
 
-path = arg.input
-img = cv2.imread(path)
-cv2.namedWindow("win",cv2.WINDOW_AUTOSIZE)
-cv2.imshow("win",img)
+filePath = arg.file
+folderPath = arg.folder
 
-cv2.waitKey(500)
-cv2.destroyAllWindows()
-print("----")
+if filePath != "default":
+	print(filePath)
+	img = cv2.imread(filePath)
+	cv2.namedWindow("win",cv2.WINDOW_AUTOSIZE)
+	cv2.imshow("win",img)
+
+	cv2.waitKey(500)
+	cv2.destroyAllWindows()
+	print("----")
+elif folderPath != "default":
+	pass
+else:
+	print("No Input !")
